se actualiaron las realciones en la entidad de inventoryproducts

+++ src/inventoryproduct/entities/inventoryproduct.entity.ts
@@ -8,7 +8,7 @@ export class Inventoryproduct {
   @PrimaryGeneratedColumn('uuid')
   id: string;
 
-  @Column('int',{
+  @Column('float',{
     default: 0
   })
   amount: number;
